requirements.txt
Flask
Flask-RESTful
Flask-JWT
Flask-SQLAlchemy
uwsgi
psycopg2
flask_cors
pandas
openpyxl
xlsxwriter